+++ tests/version-validator.sh
@@ -110,6 +110,9 @@ expect_pass '2.0.0-rc13.9'
 write_fixture '2.0.0-rc13.10' '2.0.0-rc13.10' '1.0.0'
 expect_pass '2.0.0-rc13.10'
 
+write_fixture '2.0.0-rc13.11' '2.0.0-rc13.11' '1.0.0'
+expect_pass '2.0.0-rc13.11'
+
 write_fixture '2.0.0-rc1' '2.0.0-rc1' '2.0.0'
 expect_fail '2.0.0-rc1'
 
@@ -167,8 +170,8 @@ expect_fail '2.0.0-rc13.6'
 write_fixture '2.0.0-rc13.7' '2.0.0-rc13.8' '1.0.0'
 expect_fail '2.0.0-rc13.7'
 
-write_fixture '2.0.0-rc13.9' '2.0.0-rc13.10' '1.0.0'
-expect_fail '2.0.0-rc13.9'
+write_fixture '2.0.0-rc13.10' '2.0.0-rc13.11' '1.0.0'
+expect_fail '2.0.0-rc13.10'
 
 printf 'Release version validator tests passed.\n'